+++ tests/testthat/test_model_income_tax.R
@@ -809,7 +809,7 @@ test_that("Debugger", {
   s1516 <- as.data.table(sample_file_1516_synth)
   s1516[, Med_Exp_TO_amt := 0]
   o <- model_income_tax(s1516[, Med_Exp_TO_amt := 0], "2016-17", .debug = TRUE)
-  expect_equal(names(o), c("income", "old_tax",
+  expect_equal(names(o), c("Ind", "income", "old_tax", "new_tax",
                            "base_tax.", "lito.", "lamington_offset.", "sapto.", 
                            "sbto.", "medicare_levy."))
 })
@@ -900,4 +900,52 @@ test_that("CGT (errors)", {
                regexp = "`cgt_discount_rate` was type character",
                fixed = TRUE)
 })
+
+test_that("SAPTO parameters should not go out of range (causing NAs)", {
+  skip_on_cran()
+  skip_if_not_installed("taxstats")
+  skip_on_circleci(2)
+  library(taxstats)
+  library(data.table)
+  library(hutils)
+  s1314 <- as.data.table(sample_file_1314)
+  m1314_27000 <- model_income_tax(s1314, "2016-17", sapto_lower_threshold = 27000)
+  expect_false(anyNA(m1314_27000[["new_tax"]]))
+})
+
+test_that("Issue #176", {
+  skip_if_not_installed("taxstats")
+  skip_on_circleci(3)
+  library(taxstats)
+  # Just no error
+  expect_silent(model_income_tax(sample_file_1314, 
+                                 "2016-17", 
+                                 medicare_levy_lower_family_sapto_threshold = 42000,
+                                 medicare_levy_upper_family_sapto_threshold = 52500))
+  expect_true(TRUE)
+})
+
+test_that("SAPTO modelling done for Age of entitlement report", {
+  skip_on_cran()
+  skip_if_not_installed("taxstats")
+  skip_on_circleci(2)
+  library(taxstats)
+  library(data.table)
+  library(hutils)
+  s1718_AgeOfEntitlement <-
+    project(sample_file_1314, 
+            h = 4L) %>%
+    model_income_tax("2017-18", 
+                     sapto_lower_threshold = 27e3,
+                     sapto_lower_threshold_married = 42e3,
+                     sapto_max_offset = 1160,
+                     sapto_max_offset_married = 390,
+                     medicare_levy_lower_sapto_threshold = 27000,
+                     medicare_levy_upper_sapto_threshold = 33750,
+                     medicare_levy_upper_family_threshold = 46361,
+                     medicare_levy_lower_family_sapto_threshold = 42000,
+                     medicare_levy_upper_family_sapto_threshold = 52500)
+})
+
+